feat: Add the equation based on latex and change test file
