+++ onebone/feature/__init__.py
@@ -1,5 +1,6 @@
 from .frequency import mdf, mnf, vcf
 from .gear import na4
+from .snr import snr
 from .tacho import tacho_to_angle, tacho_to_rpm
 from .time import crest_factor, kurtosis, peak2peak, rms
 
@@ -14,4 +15,5 @@ __all__ = [
     "na4",
     "tacho_to_angle",
     "tacho_to_rpm",
+    "snr",
 ]
